backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
celery[redis]==5.3.4
redis==5.0.7
openai==1.40.0
google-api-python-client==2.140.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
browser-use==0.1.8
langchain-openai==0.2.10
playwright==1.45.0
pymilvus==2.4.5
pandas==2.2.2
numpy==1.26.4
pydantic==2.8.2
python-multipart==0.0.9
aiofiles==24.1.0
httpx==0.27.0
beautifulsoup4==4.12.3
pypdf==4.3.1
python-docx==1.1.2
openpyxl==3.1.5
markdown==3.6
jinja2==3.1.4
tenacity==8.5.0
# New RAG Enhancement Dependencies
rank-bm25==0.2.2
bm25s==0.1.8  
sentence-transformers==2.2.2
torch>=2.0.0
scikit-learn==1.3.2
ragas==0.1.15
datasets==2.14.0
langchain-community==0.0.38
flashrank==0.2.0
